moved app folder into components folder and created file structure for project
