+++ client/src/components/app/App.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { Provider } from 'react-redux';
 import logo from './logo.svg';
 import './App.css';
-import Customers from './../components/Customer/customers';
-import store from './../store';
+import Customers from './../Customer/customers';
+import store from './../../store';
 
 class App extends Component {
   render () {
@@ -25,9 +25,7 @@ class App extends Component {
           </a>
         </header>
         <body>
-          <Provider store={ store }>
-            <Customers/>
-          </Provider>
+
         </body>
       </div>
     );
